LungCancerDetector/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "streamlit>=1.41.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "trafilatura>=2.0.0",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
